+++ mountpoint-s3/tests/fuse_tests/fork_test.rs
@@ -4,10 +4,13 @@
 use assert_cmd::prelude::*;
 #[cfg(not(feature = "s3express_tests"))]
 use aws_sdk_s3::primitives::ByteStream;
+use fuser::MountOption;
+use predicates::prelude::*;
 use std::fs::{self, File};
 #[cfg(not(feature = "s3express_tests"))]
 use std::io::Read;
 use std::io::{self, BufRead, BufReader, Cursor, Write};
+use std::os::fd::AsRawFd;
 use std::path::Path;
 use std::process::{Child, ExitStatus, Stdio};
 use std::time::{Duration, Instant};
@@ -16,7 +19,7 @@ use tempfile::NamedTempFile;
 use test_case::test_case;
 
 use crate::common::creds::{get_sdk_default_chain_creds, get_subsession_iam_role};
-use crate::common::fuse::read_dir_to_entry_names;
+use crate::common::fuse::{mount_for_passing_fuse_fd, read_dir_to_entry_names};
 use crate::common::s3::{
     create_objects, get_test_bucket_and_prefix, get_test_bucket_forbidden, get_test_endpoint_url, get_test_region,
     get_test_sdk_client,
@@ -24,6 +27,9 @@ use crate::common::s3::{
 use crate::common::tokio_block_on;
 #[cfg(not(feature = "s3express_tests"))]
 use crate::common::{creds::get_scoped_down_credentials, s3::get_non_test_region, s3::get_test_kms_key_id};
+
+const MOUNT_OPTION_READ_ONLY: &str = "--read-only";
+const MOUNT_OPTION_AUTO_UNMOUNT: &str = "--auto-unmount";
 
 const MAX_WAIT_DURATION: std::time::Duration = std::time::Duration::from_secs(10);
 
@@ -49,6 +55,39 @@ fn run_in_background() -> Result<(), Box<dyn std::error::Error>> {
     // verify mount status and mount entry
     assert!(exit_status.success());
     assert!(mount_exists("mountpoint-s3", mount_point.path().to_str().unwrap()));
+
+    test_read_files(&bucket, &prefix, &region, &mount_point.to_path_buf());
+
+    unmount(mount_point.path());
+
+    Ok(())
+}
+
+#[test]
+fn run_in_background_with_passed_fuse_fd() -> Result<(), Box<dyn std::error::Error>> {
+    let (bucket, prefix) = get_test_bucket_and_prefix("run_in_background_with_passed_fuse_fd");
+    let region = get_test_region();
+    let mount_point = assert_fs::TempDir::new()?;
+
+    let (fd, _mount) = mount_for_passing_fuse_fd(
+        mount_point.path(),
+        &[MountOption::FSName("mountpoint-s3-fd".to_string())],
+    );
+
+    let mut cmd = Command::cargo_bin("mount-s3")?;
+    let child = cmd
+        .arg(&bucket)
+        .arg(format!("/dev/fd/{}", fd.as_raw_fd()))
+        .arg(format!("--prefix={prefix}"))
+        .arg(format!("--region={region}"))
+        .spawn()
+        .expect("unable to spawn child");
+
+    let exit_status = wait_for_exit(child);
+
+    // verify mount status and mount entry
+    assert!(exit_status.success());
+    assert!(mount_exists("mountpoint-s3-fd", mount_point.path().to_str().unwrap()));
 
     test_read_files(&bucket, &prefix, &region, &mount_point.to_path_buf());
 
@@ -152,6 +191,74 @@ fn run_in_foreground() -> Result<(), Box<dyn std::error::Error>> {
 }
 
 #[test]
+fn run_in_foreground_with_passed_fuse_fd() -> Result<(), Box<dyn std::error::Error>> {
+    let (bucket, prefix) = get_test_bucket_and_prefix("run_in_foreground_with_passed_fuse_fd");
+    let region = get_test_region();
+    let mount_point = assert_fs::TempDir::new()?;
+
+    let (fd, _mount) = mount_for_passing_fuse_fd(
+        mount_point.path(),
+        &[MountOption::FSName("mountpoint-s3-fd".to_string())],
+    );
+
+    let mut cmd = Command::cargo_bin("mount-s3")?;
+    let mut child = cmd
+        .arg(&bucket)
+        .arg(format!("/dev/fd/{}", fd.as_raw_fd()))
+        .arg(format!("--prefix={prefix}"))
+        .arg("--foreground")
+        .arg(format!("--region={region}"))
+        .spawn()
+        .expect("unable to spawn child");
+
+    wait_for_mount("mountpoint-s3-fd", mount_point.path().to_str().unwrap());
+
+    let child_exit_status = child.try_wait().unwrap();
+    assert_eq!(
+        None, child_exit_status,
+        "child exit status should be None as it should still be running"
+    );
+
+    assert!(mount_exists("mountpoint-s3-fd", mount_point.path().to_str().unwrap()));
+
+    test_read_files(&bucket, &prefix, &region, &mount_point.to_path_buf());
+
+    unmount(mount_point.path());
+
+    Ok(())
+}
+
+#[test]
+fn run_in_background_with_passed_fuse_fd_fail_on_mount() -> Result<(), Box<dyn std::error::Error>> {
+    // the mount would fail from error 403 on HeadBucket
+    let bucket = get_test_bucket_forbidden();
+    let mount_point = assert_fs::TempDir::new()?;
+
+    let (fd, mount) = mount_for_passing_fuse_fd(
+        mount_point.path(),
+        &[MountOption::FSName("mountpoint-s3-fd".to_string())],
+    );
+
+    let mut cmd = Command::cargo_bin("mount-s3")?;
+    let child = cmd
+        .arg(&bucket)
+        .arg(format!("/dev/fd/{}", fd.as_raw_fd()))
+        .spawn()
+        .expect("unable to spawn child");
+
+    let exit_status = wait_for_exit(child);
+
+    // Drop `Mount` to trigger unmount.
+    drop(mount);
+
+    // verify mount status and mount entry
+    assert!(!exit_status.success());
+    assert!(!mount_exists("mountpoint-s3-fd", mount_point.path().to_str().unwrap()));
+
+    Ok(())
+}
+
+#[test]
 fn run_in_background_fail_on_mount() -> Result<(), Box<dyn std::error::Error>> {
     // the mount would fail from error 403 on HeadBucket
     let bucket = get_test_bucket_forbidden();
@@ -237,6 +344,107 @@ fn run_fail_on_duplicate_mount() -> Result<(), Box<dyn std::error::Error>> {
     assert!(!exit_status.success());
 
     unmount(mount_point.path());
+
+    Ok(())
+}
+
+#[test]
+fn run_fail_on_non_existent_fd() -> Result<(), Box<dyn std::error::Error>> {
+    let (bucket, prefix) = get_test_bucket_and_prefix("run_fail_on_non_existent_fd");
+    let region = get_test_region();
+
+    let mount_point = "/dev/fd/1025";
+
+    let mut cmd = Command::cargo_bin("mount-s3")?;
+    let child = cmd
+        .arg(&bucket)
+        .arg(mount_point)
+        .arg(format!("--prefix={prefix}"))
+        .arg(format!("--region={region}"))
+        .spawn()
+        .expect("unable to spawn child");
+
+    let exit_status = wait_for_exit(child);
+
+    // verify mount status
+    assert!(!exit_status.success());
+
+    // verify error message
+    let error_message = format!("mount point {} is not a valid file descriptor", mount_point);
+    cmd.assert().failure().stderr(predicate::str::contains(error_message));
+
+    Ok(())
+}
+
+#[test]
+fn run_fail_on_non_fuse_fd() -> Result<(), Box<dyn std::error::Error>> {
+    let (bucket, prefix) = get_test_bucket_and_prefix("run_fail_on_non_fuse_fd");
+    let region = get_test_region();
+
+    // 1 is fd for stdout
+    let mount_point = "/dev/fd/1";
+
+    let mut cmd = Command::cargo_bin("mount-s3")?;
+    let child = cmd
+        .arg(&bucket)
+        .arg(mount_point)
+        .arg(format!("--prefix={prefix}"))
+        .arg(format!("--region={region}"))
+        .spawn()
+        .expect("unable to spawn child");
+
+    let exit_status = wait_for_exit(child);
+
+    // verify mount status
+    assert!(!exit_status.success());
+
+    // verify error message
+    let error_message = format!(
+        "expected mount point {} to be a /dev/fuse device file descriptor but got Pipe",
+        mount_point
+    );
+    cmd.assert().failure().stderr(predicate::str::contains(error_message));
+
+    Ok(())
+}
+
+#[test_case(&[MOUNT_OPTION_READ_ONLY])]
+#[test_case(&[MOUNT_OPTION_AUTO_UNMOUNT])]
+#[test_case(&[MOUNT_OPTION_READ_ONLY, MOUNT_OPTION_AUTO_UNMOUNT])]
+fn run_fail_on_non_fuse_fd_if_mount_options_passed(mount_options: &[&str]) -> Result<(), Box<dyn std::error::Error>> {
+    let (bucket, prefix) = get_test_bucket_and_prefix("run_fail_on_non_fuse_fd_if_mount_options_passed");
+    let region = get_test_region();
+    let mount_point = assert_fs::TempDir::new()?;
+
+    let (fd, _mount) = mount_for_passing_fuse_fd(
+        mount_point.path(),
+        &[MountOption::FSName("mountpoint-s3-fd".to_string())],
+    );
+
+    let mut cmd = Command::cargo_bin("mount-s3")?;
+    cmd.arg(&bucket)
+        .arg(format!("/dev/fd/{}", fd.as_raw_fd()))
+        .arg(format!("--prefix={prefix}"))
+        .arg(format!("--region={region}"));
+
+    for opt in mount_options {
+        cmd.arg(opt);
+    }
+
+    let child = cmd.spawn().expect("unable to spawn child");
+
+    let exit_status = wait_for_exit(child);
+
+    // verify mount status
+    assert!(!exit_status.success());
+
+    // verify error message
+    let error_message = format!(
+        "Mount options: {} are ignored with FUSE fd mount point.\
+        Mount options should be passed while performing `mount` syscall in the caller process.",
+        mount_options.join(", "),
+    );
+    cmd.assert().failure().stderr(predicate::str::contains(error_message));
 
     Ok(())
 }